refactor: move content of JNI/Unsafe.hs to JNI/Unsafe/Internal.hs
This will allow to define some wrappers around JNI bindings before
exporting them (see #148)
